Refactored test helpers
1. Split into separate files
2. Renamed a few to make them more consistent
3. Deleted some redundant helpers
