+++ packages/e2e-tests/helpers/source-panel.ts
@@ -1,0 +1,280 @@
+import { expect, Locator } from "@playwright/test";
+import { waitFor } from "@playwright-testing-library/test";
+import chalk from "chalk";
+
+import { openDevToolsTab } from ".";
+import { openPauseInformationPanel } from "./pause-information-panel";
+import { openSource, openSourceExplorerPanel } from "./source-explorer-panel";
+import { Screen } from "./types";
+import { clearTextArea, debugPrint } from "./utils";
+
+export async function addBreakpoint(
+  screen: Screen,
+  options: {
+    columnIndex?: number;
+    lineNumber: number;
+    url: string;
+  }
+): Promise<void> {
+  const { lineNumber, url } = options;
+
+  debugPrint(`Adding breakpoint at ${chalk.bold(`${url}:${lineNumber}`)}`, "addBreakpoint");
+
+  await openDevToolsTab(screen);
+
+  if (url) {
+    await openSourceExplorerPanel(screen);
+    await openSource(screen, url);
+  }
+
+  const line = await getSourceLine(screen, lineNumber);
+  await line.locator(".CodeMirror-linenumber").hover();
+  await line.locator(".CodeMirror-linenumber").click();
+
+  await waitForBreakpoint(screen, options);
+}
+
+export async function addLogpoint(
+  screen: Screen,
+  options: {
+    columnIndex?: number;
+    lineNumber: number;
+    condition?: string;
+    content?: string;
+    url: string;
+  }
+): Promise<void> {
+  const { condition, content, lineNumber, url } = options;
+
+  debugPrint(`Adding log-point at ${chalk.bold(`${url}:${lineNumber}`)}`, "addLogpoint");
+
+  await openDevToolsTab(screen);
+
+  if (url) {
+    await openSourceExplorerPanel(screen);
+    await openSource(screen, url);
+  }
+
+  const line = await getSourceLine(screen, lineNumber);
+  await line.hover();
+  await line.locator('[data-test-id="ToggleLogpointButton"]').click();
+
+  await waitForLogpoint(screen, options);
+
+  if (condition || content) {
+    const line = getSourceLine(screen, lineNumber);
+    await line.locator('[data-test-name="LogpointContentSummary"]').click();
+
+    if (condition) {
+      debugPrint(`Setting log-point condition "${chalk.bold(condition)}"`, "addLogpoint");
+
+      await line.locator('[data-test-name="EditLogpointConditionButton"]').click();
+
+      // Condition and content both have text areas.
+      await expect(line.locator("textarea")).toHaveCount(2);
+
+      const textArea = await line.locator("textarea").first();
+      await textArea.focus();
+      await clearTextArea(screen, textArea);
+      await screen.keyboard.type(condition);
+    }
+
+    if (content) {
+      debugPrint(`Setting log-point content "${chalk.bold(content)}"`, "addLogpoint");
+
+      // Condition and content both have text areas.
+      // Content will always be the last one regardless of whether the condition text area is visible.
+      const textArea = await line.locator("textarea").last();
+      await textArea.waitFor();
+      await textArea.focus();
+      await clearTextArea(screen, textArea);
+      await screen.keyboard.type(content);
+    }
+
+    const saveButton = await line.locator('button[title="Save expression"]');
+    await expect(saveButton).toBeEnabled();
+    await saveButton.click();
+
+    const textArea = await line.locator("textarea");
+    await textArea.waitFor({ state: "detached" });
+  }
+}
+
+export async function closeSource(screen: Screen, url: string): Promise<void> {
+  debugPrint(`Closing source "${chalk.bold(url)}"`, "openSource");
+
+  const sourceTab = getSourceTab(screen, url);
+
+  if (await sourceTab.isVisible()) {
+    await sourceTab.locator("button").click();
+  }
+
+  await sourceTab.waitFor({ state: "detached" });
+}
+
+export async function getSelectedLineNumber(screen: Screen): Promise<number | null> {
+  const lineNumber = await screen.locator(".new-debug-line .CodeMirror-linenumber");
+  const textContent = await lineNumber.textContent();
+  return textContent !== null ? parseInt(textContent, 10) : null;
+}
+
+export function getSourceLine(screen: Screen, lineNumber: number): Locator {
+  return screen
+    .locator(".CodeMirror-code div", {
+      has: screen.locator(`.CodeMirror-linenumber:text-is("${lineNumber}")`),
+    })
+    .first();
+}
+
+export function getSourceTab(screen: Screen, url: string): Locator {
+  return screen.locator(`[data-test-name="Source-${url}"]`);
+}
+
+export async function removeAllBreakpoints(screen: Screen): Promise<void> {
+  debugPrint(`Removing all breakpoints for the current source`, "removeBreakpoint");
+
+  while (true) {
+    const breakpoint = screen.locator(".editor.new-breakpoint");
+    const count = await breakpoint.count();
+    if (count > 0) {
+      await breakpoint.click();
+    } else {
+      return;
+    }
+  }
+}
+
+export async function removeBreakpoint(
+  screen: Screen,
+  options: {
+    lineNumber: number;
+    url: string;
+  }
+): Promise<void> {
+  const { lineNumber, url } = options;
+
+  debugPrint(`Removing breakpoint at ${chalk.bold(`${url}:${lineNumber}`)}`, "removeBreakpoint");
+
+  await openDevToolsTab(screen);
+
+  if (url) {
+    await openSourceExplorerPanel(screen);
+    await openSource(screen, url);
+  }
+
+  const line = await getSourceLine(screen, lineNumber);
+  await line.locator(".CodeMirror-linenumber").click({ force: true });
+}
+
+export async function waitForBreakpoint(
+  screen: Screen,
+  options: {
+    columnIndex?: number;
+    lineNumber: number;
+    url: string;
+  }
+): Promise<void> {
+  const { columnIndex, lineNumber, url } = options;
+
+  debugPrint(
+    `Waiting for breakpoint at ${chalk.bold(`${url}:${lineNumber}`)}`,
+    "waitForBreakpoint"
+  );
+
+  await openPauseInformationPanel(screen);
+
+  const breakpointGroup = await screen.waitForSelector(
+    `.breakpoints-list-source:has-text("${url}")`
+  );
+
+  if (columnIndex != null) {
+    await breakpointGroup.waitForSelector(
+      `.breakpoint-line:has-text("${lineNumber}:${columnIndex}")`
+    );
+  } else {
+    await breakpointGroup.waitForSelector(`.breakpoint-line:has-text("${lineNumber}")`);
+  }
+}
+
+export async function waitForLogpoint(
+  screen: Screen,
+  options: {
+    columnIndex?: number;
+    lineNumber: number;
+    url: string;
+  }
+): Promise<void> {
+  const { lineNumber, url } = options;
+
+  debugPrint(`Waiting for log-point at ${chalk.bold(`${url}:${lineNumber}`)}`, "waitForLogpoint");
+
+  await openDevToolsTab(screen);
+
+  if (url) {
+    await openSourceExplorerPanel(screen);
+    await openSource(screen, url);
+  }
+
+  const line = getSourceLine(screen, lineNumber);
+  await line.locator(".CodeMirror-linewidget").waitFor();
+}
+
+export async function verifyLogpointStep(
+  screen: Screen,
+  expectedStatus: string,
+  options: {
+    lineNumber: number;
+    url?: string;
+  }
+): Promise<void> {
+  const { lineNumber, url } = options;
+
+  await openSourceExplorerPanel(screen);
+
+  if (url) {
+    await openSource(screen, url);
+  }
+
+  debugPrint(
+    `Verifying breakpoint status "${chalk.bold(expectedStatus)}" for line ${chalk.bold(
+      options.lineNumber
+    )}`,
+    "verifyLogpointStep"
+  );
+
+  const line = await getSourceLine(screen, lineNumber);
+  const status = line.locator(
+    `[data-test-name="LogpointPanel-BreakpointStatus"]:has-text("${expectedStatus}")`
+  );
+  await status.waitFor({ state: "visible" });
+}
+
+// TODO [FE-626] Rewrite this helper to reduce complexity.
+export async function waitForSelectedSource(screen: Screen, url: string) {
+  return waitFor(async () => {
+    const editorPanel = screen.locator("#toolbox-content-debugger");
+    const sourceHeader = editorPanel.locator(`[data-test-name="Source-${url}"]`);
+    const isTabActive = (await sourceHeader.getAttribute("data-status")) === "active";
+
+    // HACK Assume that the source file has loaded when the combined text of the first
+    // 10 lines is no longer an empty string
+    const codeMirrorLines = editorPanel.locator(".CodeMirror-code .CodeMirror-line");
+
+    const numLines = await codeMirrorLines.count();
+    // Could be fewer than 10 lines visible, and I can't find a `.slice()` on `Locator`
+    const linesToGrab = Math.min(numLines, 10);
+
+    const lineTextPromises = Array.from({ length: linesToGrab }).map((_, i) => {
+      return codeMirrorLines.nth(i).textContent();
+    });
+    const lineTexts = await Promise.all(lineTextPromises);
+
+    const combinedLineText = lineTexts
+      .join()
+      .trim()
+      // Remove zero-width spaces, which would be considered non-empty
+      .replace(/[\u200B-\u200D\uFEFF]/g, "");
+
+    expect(isTabActive && lineTexts.length > 0 && combinedLineText !== "").toBe(true);
+  });
+}
