+++ packages/e2e-tests/tests/breakpoints-01.test.ts
@@ -1,37 +1,34 @@
-import {
-  addBreakpoint,
-  checkEvaluateInTopFrame,
-  clickDevTools,
-  openExample,
-  test,
-  resumeToLine,
-  rewindToLine,
-} from "../helpers";
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { executeAndVerifyTerminalExpression } from "../helpers/console-panel";
+import { resumeToLine, rewindToLine } from "../helpers/pause-information-panel";
+import { addBreakpoint } from "../helpers/source-panel";
+
+const url = "doc_rr_basic.html";
 
 test(`Test basic breakpoint functionality.`, async ({ screen }) => {
-  await openExample(screen, "doc_rr_basic.html");
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
-  await addBreakpoint(screen, { lineNumber: 21, url: "doc_rr_basic.html" });
+  await addBreakpoint(screen, { lineNumber: 21, url });
 
   await rewindToLine(screen, {
     lineNumber: 21,
   });
-  await checkEvaluateInTopFrame(screen, "number", "10");
+  await executeAndVerifyTerminalExpression(screen, "number", "10");
   await rewindToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "9");
+  await executeAndVerifyTerminalExpression(screen, "number", "9");
   await rewindToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "8");
+  await executeAndVerifyTerminalExpression(screen, "number", "8");
   await rewindToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "7");
+  await executeAndVerifyTerminalExpression(screen, "number", "7");
   await rewindToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "6");
+  await executeAndVerifyTerminalExpression(screen, "number", "6");
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "7");
+  await executeAndVerifyTerminalExpression(screen, "number", "7");
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "8");
+  await executeAndVerifyTerminalExpression(screen, "number", "8");
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "9");
+  await executeAndVerifyTerminalExpression(screen, "number", "9");
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "10");
+  await executeAndVerifyTerminalExpression(screen, "number", "10");
 });
