+++ packages/e2e-tests/tests/breakpoints-03.test.ts
@@ -1,31 +1,27 @@
-import {
-  addBreakpoint,
-  checkEvaluateInTopFrame,
-  clickDevTools,
-  openExample,
-  removeBreakpoint,
-  resumeToLine,
-  rewindToLine,
-  test,
-} from "../helpers";
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { executeAndVerifyTerminalExpression } from "../helpers/console-panel";
+import { resumeToLine, rewindToLine } from "../helpers/pause-information-panel";
+import { addBreakpoint, removeBreakpoint } from "../helpers/source-panel";
+
+const url = "doc_rr_basic.html";
 
 test(`Test stepping forward through breakpoints when rewound before the first one.`, async ({
   screen,
 }) => {
-  await openExample(screen, "doc_rr_basic.html");
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
-  await addBreakpoint(screen, { lineNumber: 8, url: "doc_rr_basic.html" });
+  await addBreakpoint(screen, { lineNumber: 8, url });
   // Rewind to when the point was hit
   await rewindToLine(screen, { lineNumber: 8 });
   // Rewind further (past the first hit)
   await rewindToLine(screen);
 
-  await removeBreakpoint(screen, { lineNumber: 8, url: "doc_rr_basic.html" });
+  await removeBreakpoint(screen, { lineNumber: 8, url });
 
-  await addBreakpoint(screen, { lineNumber: 21, url: "doc_rr_basic.html" });
+  await addBreakpoint(screen, { lineNumber: 21, url });
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "1");
+  await executeAndVerifyTerminalExpression(screen, "number", "1");
   await resumeToLine(screen, { lineNumber: 21 });
-  await checkEvaluateInTopFrame(screen, "number", "2");
+  await executeAndVerifyTerminalExpression(screen, "number", "2");
 });
