+++ packages/e2e-tests/tests/breakpoints-04.test.ts
@@ -1,17 +1,14 @@
-import {
-  addBreakpoint,
-  clickDevTools,
-  openExample,
-  rewindToLine,
-  resumeToLine,
-  Screen,
-  test,
-} from "../helpers";
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { resumeToLine, rewindToLine } from "../helpers/pause-information-panel";
+import { addBreakpoint } from "../helpers/source-panel";
+import { Screen } from "../helpers/types";
+
+const url = "doc_control_flow.html";
 
 // Test hitting breakpoints when using tricky control flow constructs:
 test(`catch, finally, generators, and async/await.`, async ({ screen }) => {
-  await openExample(screen, "doc_control_flow.html");
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
   await rewindToBreakpoint(screen, 10);
   await resumeToBreakpoint(screen, 12);
@@ -29,12 +26,12 @@ test(`catch, finally, generators, and async/await.`, async ({ screen }) => {
   await resumeToBreakpoint(screen, 72);
 
   async function rewindToBreakpoint(screen: Screen, lineNumber: number) {
-    await addBreakpoint(screen, { lineNumber, url: "doc_control_flow.html" });
+    await addBreakpoint(screen, { lineNumber, url });
     await rewindToLine(screen, { lineNumber });
   }
 
   async function resumeToBreakpoint(screen: Screen, lineNumber: number) {
-    await addBreakpoint(screen, { lineNumber, url: "doc_control_flow.html" });
+    await addBreakpoint(screen, { lineNumber, url });
     await resumeToLine(screen, { lineNumber });
   }
 });
