+++ packages/e2e-tests/tests/breakpoints-05.test.ts
@@ -1,18 +1,17 @@
+import { openDevToolsTab, startTest, test } from "../helpers";
 import {
-  addBreakpoint,
-  clickDevTools,
-  openExample,
-  openPauseInformation,
-  removeBreakpoint,
+  openPauseInformationPanel,
   resumeToLine,
   rewindToLine,
-  test,
-} from "../helpers";
+} from "../helpers/pause-information-panel";
+import { addBreakpoint, removeBreakpoint } from "../helpers/source-panel";
+
+const url = "doc_debugger_statements.html";
 
 test(`Test interaction of breakpoints with debugger statements.`, async ({ screen }) => {
-  await openExample(screen, "doc_debugger_statements.html");
-  await clickDevTools(screen);
-  await openPauseInformation(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
+  await openPauseInformationPanel(screen);
 
   // Without any breakpoints, this test should rewind to the closest debugger statement.
   await rewindToLine(screen, { lineNumber: 9 });
@@ -20,7 +19,7 @@ test(`Test interaction of breakpoints with debugger statements.`, async ({ scree
   // Without a breakpoints being the next nearest thing, we should rewind to it.
   await addBreakpoint(screen, {
     lineNumber: 8,
-    url: "doc_debugger_statements.html",
+    url,
   });
   await rewindToLine(screen, { lineNumber: 8 });
   await resumeToLine(screen, { lineNumber: 9 });
@@ -28,7 +27,7 @@ test(`Test interaction of breakpoints with debugger statements.`, async ({ scree
   // Without any breakpoints (again), we should rewind to debugger statements.
   await removeBreakpoint(screen, {
     lineNumber: 8,
-    url: "doc_debugger_statements.html",
+    url,
   });
   await rewindToLine(screen, { lineNumber: 7 });
   await resumeToLine(screen, { lineNumber: 9 });
