+++ packages/e2e-tests/tests/breakpoints-06.test.ts
@@ -1,22 +1,19 @@
 import { expect } from "@playwright/test";
-import {
-  addLogpoint,
-  clickDevTools,
-  getConsoleMessage,
-  openExample,
-  Screen,
-  test,
-} from "../helpers";
+
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { findConsoleMessage } from "../helpers/console-panel";
+import { addLogpoint } from "../helpers/source-panel";
+import { Screen } from "../helpers/types";
 
 async function checkMessageLocation(screen: Screen, text: string, location: string) {
-  const message = await getConsoleMessage(screen, text, "log-point");
+  const message = await findConsoleMessage(screen, text, "log-point");
   const textContent = await message.textContent();
   expect(textContent!.includes(location)).toBeTruthy();
 }
 
 test(`Test log point in a sourcemapped file.`, async ({ screen }) => {
-  await openExample(screen, "doc_prod_bundle.html");
-  await clickDevTools(screen);
+  await startTest(screen, "doc_prod_bundle.html");
+  await openDevToolsTab(screen);
 
   // Log point added to line 15 should map to line 15
   await addLogpoint(screen, { lineNumber: 15, url: "bundle_input.js" });
