+++ packages/e2e-tests/tests/inspector-03.test.ts
@@ -1,32 +1,29 @@
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { openConsolePanel, warpToMessage } from "../helpers/console-panel";
 import {
-  test,
-  openExample,
-  clickDevTools,
-  selectConsole,
-  warpToMessage,
-  selectInspector,
-  addBreakpoint,
-  rewindToLine,
+  activateInspectorTool,
   checkComputedStyle,
-  selectMarkupNode,
-} from "../helpers";
+  selectElementsRowWithText,
+} from "../helpers/elements-panel";
+import { rewindToLine } from "../helpers/pause-information-panel";
+import { addBreakpoint } from "../helpers/source-panel";
 
 test("Test that styles for elements can be viewed.", async ({ screen }) => {
-  await openExample(screen, "doc_inspector_styles.html");
-  await clickDevTools(screen);
-  await selectConsole(screen);
+  await startTest(screen, "doc_inspector_styles.html");
+  await openDevToolsTab(screen);
+  await openConsolePanel(screen);
   await warpToMessage(screen, "ExampleFinished");
-  await selectInspector(screen);
+  await activateInspectorTool(screen);
 
-  await selectMarkupNode(screen, "body");
+  await selectElementsRowWithText(screen, "body");
   await checkComputedStyle(screen, "background-color", "rgb(0, 128, 0)");
 
-  await selectMarkupNode(screen, "maindiv");
+  await selectElementsRowWithText(screen, "maindiv");
   await checkComputedStyle(screen, "background-color", "rgb(0, 0, 255)");
 
   await addBreakpoint(screen, { url: "doc_inspector_styles.html", lineNumber: 11 });
   await rewindToLine(screen, { lineNumber: 11 });
 
-  await selectMarkupNode(screen, "maindiv");
+  await selectElementsRowWithText(screen, "maindiv");
   await checkComputedStyle(screen, "background-color", "rgb(255, 0, 0)");
 });
