+++ packages/e2e-tests/tests/logpoints-01.test.ts
@@ -1,23 +1,20 @@
 import { expect } from "@playwright/test";
 
+import { openDevToolsTab, startTest, test } from "../helpers";
 import {
-  addLogpoint,
-  checkEvaluateInTopFrame,
-  clickDevTools,
-  openExample,
-  test,
-  resumeToLine,
-  selectConsole,
-  getConsoleMessage,
+  executeAndVerifyTerminalExpression,
+  findConsoleMessage,
+  openConsolePanel,
   seekToConsoleMessage,
-  reverseStepOverToLine,
-} from "../helpers";
+} from "../helpers/console-panel";
+import { resumeToLine, reverseStepOverToLine } from "../helpers/pause-information-panel";
+import { addLogpoint } from "../helpers/source-panel";
 
 const url = "doc_rr_basic.html";
 
 test(`log-points appear in the correct order and allow time warping`, async ({ screen }) => {
-  await openExample(screen, url);
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
   await addLogpoint(screen, {
     content: '"Logpoint Number " + number',
@@ -37,20 +34,20 @@ test(`log-points appear in the correct order and allow time warping`, async ({ s
     url,
   });
 
-  await selectConsole(screen);
+  await openConsolePanel(screen);
 
-  const loadingMessages = await getConsoleMessage(screen, "Loading", "log-point");
+  const loadingMessages = await findConsoleMessage(screen, "Loading", "log-point");
   await expect(loadingMessages).toHaveCount(0);
 
-  const logPointMessages = await getConsoleMessage(screen, "Logpoint", "log-point");
+  const logPointMessages = await findConsoleMessage(screen, "Logpoint", "log-point");
   await expect(logPointMessages).toHaveCount(12);
   await expect(logPointMessages.first()).toHaveText(/Beginning/);
   await expect(logPointMessages.last()).toHaveText(/Ending/);
 
-  const message = await getConsoleMessage(screen, "Number 5", "log-point");
+  const message = await findConsoleMessage(screen, "Number 5", "log-point");
   await seekToConsoleMessage(screen, message);
 
-  await checkEvaluateInTopFrame(screen, "number", 5);
+  await executeAndVerifyTerminalExpression(screen, "number", 5);
   await reverseStepOverToLine(screen, 19);
 
   // The log-point acts like a breakpoint when resuming.
