+++ packages/e2e-tests/tests/logpoints-03.test.ts
@@ -1,22 +1,17 @@
 import { expect } from "@playwright/test";
 
-import {
-  addEventListenerLogpoints,
-  clickDevTools,
-  getConsoleMessage,
-  openExample,
-  test,
-} from "../helpers";
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { addEventListenerLogpoints, findConsoleMessage } from "../helpers/console-panel";
 
 const url = "doc_events.html";
 
 test(`should display event properties in the console`, async ({ screen }) => {
-  await openExample(screen, url);
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
   await addEventListenerLogpoints(screen, ["event.mouse.click"]);
 
-  const message = await getConsoleMessage(screen, "MouseEvent", "event");
+  const message = await findConsoleMessage(screen, "MouseEvent", "event");
 
   await expect(message).toContainText('type: "click"');
   await expect(message).toContainText("target: <div");
