+++ packages/e2e-tests/tests/logpoints-04.test.ts
@@ -1,37 +1,34 @@
 import { expect } from "@playwright/test";
 
+import { openDevToolsTab, startTest, test } from "../helpers";
 import {
-  checkEvaluateInTopFrame,
-  clickDevTools,
-  getConsoleMessage,
-  openExample,
-  reverseStepOverToLine,
+  executeAndVerifyTerminalExpression,
+  findConsoleMessage,
+  openConsolePanel,
   seekToConsoleMessage,
-  selectConsole,
-  test,
-  waitForFrameTimeline,
-} from "../helpers";
+} from "../helpers/console-panel";
+import { reverseStepOverToLine, waitForFrameTimeline } from "../helpers/pause-information-panel";
 
 const url = "doc_exceptions.html";
 
 test(`should display exceptions in the console`, async ({ screen }) => {
-  await openExample(screen, url);
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
-  await selectConsole(screen);
+  await openConsolePanel(screen);
 
   await screen.queryByTestId("FilterToggle-exceptions").click();
 
-  let messages = await getConsoleMessage(screen, undefined, "exception");
+  let messages = await findConsoleMessage(screen, undefined, "exception");
   await expect(messages).toHaveCount(20);
 
-  messages = await getConsoleMessage(screen, "number: 4", "exception");
+  messages = await findConsoleMessage(screen, "number: 4", "exception");
   await seekToConsoleMessage(screen, messages.first());
   await waitForFrameTimeline(screen, "100%");
 
-  messages = await getConsoleMessage(screen, "number: 10", "exception");
+  messages = await findConsoleMessage(screen, "number: 10", "exception");
   await seekToConsoleMessage(screen, messages.first());
-  await checkEvaluateInTopFrame(screen, "number * 10", "40");
+  await executeAndVerifyTerminalExpression(screen, "number * 10", "40");
 
   await reverseStepOverToLine(screen, 15);
   await waitForFrameTimeline(screen, "0%");
