+++ packages/e2e-tests/tests/object_preview-01-chromium.test.ts
@@ -1,6 +1,6 @@
 // // Test the objects produced by console.log() calls and by evaluating various
 // test(`expressions in the console after time warping.`, async ({ screen }) => {
-//   await Test.selectConsole();
+//   await Test.openConsolePanel();
 
 //   // Several objects currently show up differently in chromium.
 //   const target = await Test.getRecordingTarget();
@@ -40,12 +40,12 @@
 
 //   await Test.warpToMessage("Done");
 
-//   await Test.executeInConsole("Error('helo')");
+//   await Test.executeTerminalExpression("Error('helo')");
 //   await Test.waitForMessage('Error: "helo"');
 
 //   // Defining a new function like this doesn't currently work in chromium.
 //   if (target == "gecko") {
-//     await Test.executeInConsole(`
+//     await Test.executeTerminalExpression(`
 //           function f() {
 //             throw Error("there");
 //           }
@@ -55,15 +55,15 @@
 //     await Test.waitForMessage('Error: "there"');
 //   }
 
-//   Test.executeInConsole("Array(1, 2, 3)");
+//   Test.executeTerminalExpression("Array(1, 2, 3)");
 //   msg = await Test.waitForMessage("Array(3) [ 1, 2, 3 ]");
 //   await Test.checkMessageObjectContents(msg, ["0: 1", "1: 2", "2: 3", "length: 3"]);
 
-//   await Test.executeInConsole("new Uint8Array([1, 2, 3, 4])");
+//   await Test.executeTerminalExpression("new Uint8Array([1, 2, 3, 4])");
 //   msg = await Test.waitForMessage("Uint8Array(4) [ 1, 2, 3, 4 ]");
 //   await Test.checkMessageObjectContents(msg, ["0: 1", "1: 2", "2: 3", "3: 4", "length: 4"]);
 
-//   await Test.executeInConsole(`RegExp("abd", "g")`);
+//   await Test.executeTerminalExpression(`RegExp("abd", "g")`);
 //   msg = await Test.waitForMessage("/abd/g");
 
 //   // RegExp object properties are not currently available in chromium.
@@ -71,11 +71,11 @@
 //     await Test.checkMessageObjectContents(msg, ["global: true", `source: "abd"`]);
 //   }
 
-//   await Test.executeInConsole("new Set([1, 2, 3])");
+//   await Test.executeTerminalExpression("new Set([1, 2, 3])");
 //   msg = await Test.waitForMessage("Set(3) [ 1, 2, 3 ]");
 //   await Test.checkMessageObjectContents(msg, ["0: 1", "1: 2", "2: 3", "size: 3"], ["<entries>"]);
 
-//   await Test.executeInConsole("new Map([[1, {a:1}], [2, {b:2}]])");
+//   await Test.executeTerminalExpression("new Map([[1, {a:1}], [2, {b:2}]])");
 //   msg = await Test.waitForMessage("Map { 1 → {…}, 2 → {…} }");
 //   await Test.checkMessageObjectContents(
 //     msg,
@@ -83,11 +83,11 @@
 //     ["<entries>"]
 //   );
 
-//   await Test.executeInConsole("new WeakSet([{a:1}, {b:2}])");
+//   await Test.executeTerminalExpression("new WeakSet([{a:1}, {b:2}])");
 //   msg = await Test.waitForMessage("WeakSet(2) [ {…}, {…} ]");
 //   await Test.checkMessageObjectContents(msg, ["Object { a: 1 }", "Object { b: 2 }"], ["<entries>"]);
 
-//   await Test.executeInConsole("new WeakMap([[{a:1},{b:1}], [{a:2},{b:2}]])");
+//   await Test.executeTerminalExpression("new WeakMap([[{a:1},{b:1}], [{a:2},{b:2}]])");
 //   msg = await Test.waitForMessage("WeakMap { {…} → {…}, {…} → {…} }");
 //   await Test.checkMessageObjectContents(
 //     msg,
@@ -95,7 +95,7 @@
 //     ["<entries>"]
 //   );
 
-//   await Test.executeInConsole("new Promise(() => {})");
+//   await Test.executeTerminalExpression("new Promise(() => {})");
 //   msg = await Test.waitForMessage("Promise {  }");
 
 //   // Promise contents aren't currently available in chromium.
@@ -103,21 +103,21 @@
 //     await Test.checkMessageObjectContents(msg, ['"pending"'], []);
 //   }
 
-//   await Test.executeInConsole("Promise.resolve({ a: 1 })");
+//   await Test.executeTerminalExpression("Promise.resolve({ a: 1 })");
 //   msg = await Test.waitForMessage("Promise {  }");
 
 //   if (target == "gecko") {
 //     await Test.checkMessageObjectContents(msg, ['"fulfilled"', "a: 1"], ["<value>"]);
 //   }
 
-//   await Test.executeInConsole("Promise.reject({ a: 1 })");
+//   await Test.executeTerminalExpression("Promise.reject({ a: 1 })");
 //   msg = await Test.waitForMessage("Promise {  }");
 
 //   if (target == "gecko") {
 //     await Test.checkMessageObjectContents(msg, ['"rejected"', "a: 1"], ["<value>"]);
 //   }
 
-//   await Test.executeInConsole("baz");
+//   await Test.executeTerminalExpression("baz");
 //   msg = await Test.waitForMessage("function baz()");
 //   Test.checkJumpIcon(msg);
 // });
