+++ packages/e2e-tests/tests/scopes_rerender.test.ts
@@ -1,18 +1,15 @@
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { openConsolePanel, warpToMessage } from "../helpers/console-panel";
 import {
-  test,
-  openExample,
-  clickDevTools,
-  selectConsole,
-  warpToMessage,
-  waitForScopeValue,
-  selectFrame,
   expandFirstScope,
-} from "../helpers";
+  selectFrame,
+  waitForScopeValue,
+} from "../helpers/pause-information-panel";
 
 test("Test that scopes are rerendered.", async ({ screen }) => {
-  await openExample(screen, "doc_recursion.html");
-  await clickDevTools(screen);
-  await selectConsole(screen);
+  await startTest(screen, "doc_recursion.html");
+  await openDevToolsTab(screen);
+  await openConsolePanel(screen);
 
   await warpToMessage(screen, "Hello 1");
   await expandFirstScope(screen);
