+++ packages/e2e-tests/tests/stepping-01.test.ts
@@ -1,41 +1,32 @@
-import {
-  test,
-  checkEvaluateInTopFrame,
-  openExample,
-  rewind,
-  rewindToLine,
-  reverseStepOver,
-  stepOver,
-  clickSourceTreeNode,
-  clickDevTools,
-  toggleBreakpoint,
-  togglePausePane,
-  openPauseInformation,
-} from "../helpers";
+import { openDevToolsTab, startTest, test } from "../helpers";
+import { executeAndVerifyTerminalExpression } from "../helpers/console-panel";
+import { reverseStepOver, rewindToLine, stepOver } from "../helpers/pause-information-panel";
+import { clickSourceTreeNode } from "../helpers/source-explorer-panel";
+import { addBreakpoint } from "../helpers/source-panel";
+
+const url = "doc_rr_basic.html";
 
 test("Test basic step-over/back functionality.", async ({ screen }) => {
-  await openExample(screen, "doc_rr_basic.html");
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
   // Open doc_rr_basic.html
   await clickSourceTreeNode(screen, "test");
   await clickSourceTreeNode(screen, "examples");
-  await clickSourceTreeNode(screen, "doc_rr_basic.html");
+  await clickSourceTreeNode(screen, url);
 
   // Pause on line 20
-  await toggleBreakpoint(screen, 20);
+  await addBreakpoint(screen, { lineNumber: 20, url });
   await rewindToLine(screen);
 
   // Should get ten when evaluating number.
-  await checkEvaluateInTopFrame(screen, "number", "10");
+  await executeAndVerifyTerminalExpression(screen, "number", "10");
 
   // Should get nine when stepping over.
   await reverseStepOver(screen);
-  await checkEvaluateInTopFrame(screen, "number", "9");
+  await executeAndVerifyTerminalExpression(screen, "number", "9");
 
   // Should get ten when stepping over.
   await stepOver(screen);
-  await checkEvaluateInTopFrame(screen, "number", "10");
-
-  await new Promise(r => setTimeout(r, 1_000));
+  await executeAndVerifyTerminalExpression(screen, "number", "10");
 });
