+++ packages/e2e-tests/tests/stepping-05.test.ts
@@ -1,35 +1,36 @@
+import { openDevToolsTab, startTest, test } from "../helpers";
 import {
-  test,
-  openExample,
-  clickDevTools,
-  rewindToLine,
-  stepOverToLine,
-  stepOutToLine,
-  stepInToLine,
   addEventListenerLogpoints,
-  addBreakpoint,
-  resumeToLine,
-  selectConsole,
+  openConsolePanel,
   warpToMessage,
-  delay,
-} from "../helpers";
+} from "../helpers/console-panel";
+import {
+  resumeToLine,
+  rewindToLine,
+  stepInToLine,
+  stepOutToLine,
+  stepOverToLine,
+} from "../helpers/pause-information-panel";
+import { addBreakpoint } from "../helpers/source-panel";
+
+const url = "doc_minified.html";
 
 test(`Test stepping in pretty-printed code`, async ({ screen }) => {
   screen.setDefaultTimeout(120000);
-  await openExample(screen, "doc_minified.html");
-  await clickDevTools(screen);
+  await startTest(screen, url);
+  await openDevToolsTab(screen);
 
   await addBreakpoint(screen, { url: "bundle_input.js", lineNumber: 4 });
   await rewindToLine(screen, { lineNumber: 4 });
   await stepInToLine(screen, 1);
 
   // Add a breakpoint in minified.html and resume to there
-  await addBreakpoint(screen, { url: "doc_minified.html", lineNumber: 8 });
+  await addBreakpoint(screen, { url, lineNumber: 8 });
   await resumeToLine(screen, { lineNumber: 8 });
   await stepOverToLine(screen, 8);
   await stepOverToLine(screen, 9);
 
-  await selectConsole(screen);
+  await openConsolePanel(screen);
   await addEventListenerLogpoints(screen, ["event.mouse.click"]);
   await warpToMessage(screen, "click", 14);
