+++ packages/markerikson-stack-client-prototype/src/features/sources/sourcesCache.ts
@@ -2,7 +2,7 @@ import { unstable_getCacheForType as getCacheForType } from "react";
 import type { AppStore } from "../../app/store";
 
 import { api, SourceGroups } from "../../app/api";
-import { selectSourceDetails, SourceDetails } from "./sourcesSlice";
+import { openSourceDetails, SourceDetails } from "./sourcesSlice";
 
 /*
  * This entire file is _mostly_ WIP PROOF OF CONCEPT!
@@ -59,7 +59,7 @@ export function getSourceDetails(store: AppStore): SourceDetails[] {
   if (!hasPromiseListenerBeenAttached) {
     hasPromiseListenerBeenAttached = true;
     promise.then(result => {
-      sourceDetails = selectSourceDetails(store.getState());
+      sourceDetails = openSourceDetails(store.getState());
     });
   }
